Add files via upload
Each file is showing the code used for analysis of the data and results generation of the article mentioned. It is code made specifically for this data set, and applying the code to a different set might require edition of its original form. Nonetheless, I hope it can be useful for people starting to analyze scRNA seq data.